requirements.txt
Flask==3.0.0
tensorflow==2.18.0
numpy==1.26.3
Pillow==10.1.0
Werkzeug==3.0.1